追加：is_sametime()のテストと実装を追加。テストOK。 時間の計算はすべてdoubleで実施している。UnixTimeをdoubleにしているため桁不足にならないか注意する

+++ include/cugo_ros2_control2/cugo.hpp
@@ -10,15 +10,15 @@ struct RPM
 
 struct Twist
 {
-  float linear_x;
-  float angular_z;
+  double linear_x;
+  double angular_z;
 };
 
 struct Odom
 {
-  float x;
-  float y;
-  float yaw;
+  double x;
+  double y;
+  double yaw;
 };
 
 
@@ -30,33 +30,33 @@ class CuGo
 public:
   CuGo();
   CuGo(
-    float config_l_radius, float config_r_radius, float config_tread,
-    float config_reduction_ratio, int config_encoder_resolution);
+    double config_l_radius, double config_r_radius, double config_tread,
+    double config_reduction_ratio, int config_encoder_resolution);
   void set_params();
-  RPM calc_rpm(float linear_x, float angular_z);
-  Twist calc_twist(int count_diff_l, int count_diff_r, float dt);
-  Odom calc_odom(Odom odom, Twist twist, float dt);
+  RPM calc_rpm(double linear_x, double angular_z);
+  Twist calc_twist(int count_diff_l, int count_diff_r, double dt);
+  Odom calc_odom(Odom odom, Twist twist, double dt);
   bool check_invalid_value();
   bool check_timeout();
   void initialize();
   int get_mcu_init_value();
-  float get_tread();
-  float get_l_wheel_radius();
-  float get_r_wheel_radius();
-  float get_reduction_ratio();
-  float get_encoder_resolution();
-  void set_twist(float linear_x, float angular_z);
-  void set_difftime_cmdvel(float difftime);
+  double get_tread();
+  double get_l_wheel_radius();
+  double get_r_wheel_radius();
+  double get_reduction_ratio();
+  double get_encoder_resolution();
+  void set_twist(double linear_x, double angular_z);
+  void set_difftime_cmdvel(double difftime);
 
 private:
-  float tread;
-  float l_wheel_radius, r_wheel_radius;
-  float reduction_ratio;
+  double tread;
+  double l_wheel_radius, r_wheel_radius;
+  double reduction_ratio;
   int encoder_resolution;
   int count_diff_l, count_diff_r;
   bool serial_timeout, twist_timeout;
   Twist recv_twist;
-  float difftime_cmdvel;
+  double difftime_cmdvel;
 };
 
 } // namespace cugo_ros2_control2
